+++ guillotina/content.py
@@ -29,6 +29,7 @@ from guillotina.exceptions import ConflictIdOnContainer
 from guillotina.exceptions import InvalidContentType
 from guillotina.exceptions import NoPermissionToAdd
 from guillotina.exceptions import NotAllowedContentType
+from guillotina.exceptions import NotAllowedParentContentType
 from guillotina.exceptions import PreconditionFailed
 from guillotina.exceptions import TransactionNotFound
 from guillotina.interfaces import DEFAULT_ADD_PERMISSION
@@ -36,6 +37,7 @@ from guillotina.interfaces import IAddons
 from guillotina.interfaces import IAnnotations
 from guillotina.interfaces import IAsyncBehavior
 from guillotina.interfaces import IBehavior
+from guillotina.interfaces import IConstrainParentTypes
 from guillotina.interfaces import IConstrainTypes
 from guillotina.interfaces import IContainer
 from guillotina.interfaces import IFolder
@@ -102,6 +104,7 @@ class ResourceFactory(Factory):
         behaviors=None,
         add_permission=DEFAULT_ADD_PERMISSION,
         allowed_types=None,
+        allowed_parent_types=None,
     ):
         super(ResourceFactory, self).__init__(
             klass, title, description, tuple(filter(bool, [schema] + list(behaviors) or list()))
@@ -111,6 +114,7 @@ class ResourceFactory(Factory):
         self.behaviors = behaviors or ()
         self.add_permission = add_permission
         self.allowed_types = allowed_types
+        self.allowed_parent_types = allowed_parent_types
 
     @profilable
     def __call__(self, id, parent=None, *args, **kw):
@@ -604,6 +608,11 @@ async def create_content_in_container(
             # the factory sets id
             continue
         setattr(obj, key, value)
+
+    parent_type_constrains = IConstrainParentTypes(obj, None)
+    if parent_type_constrains is not None:
+        if not parent_type_constrains.is_type_allowed(parent.type_name):
+            raise NotAllowedParentContentType(str(obj), parent.type_name)
 
     txn: Optional[ITransaction]
     if hasattr(parent, "_get_transaction"):