+++ guillotina/exceptions.py
@@ -40,6 +40,18 @@ class NotAllowedContentType(Exception):
         )
 
 
+class NotAllowedParentContentType(Exception):
+    def __init__(self, content, parent_content_type):
+        super().__init__()
+        self.content = content
+        self.parent_content_type = parent_content_type
+
+    def __repr__(self):
+        return "Not allowed {content} on {parent_content_type}".format(
+            content=self.content, parent_content_type=self.parent_content_type
+        )
+
+
 class ConflictIdOnContainer(Exception):
     def __init__(self, pg_exc):
         super().__init__()