+++ guillotina/_settings.py
@@ -55,7 +55,7 @@ app_settings: Dict[str, Any] = {
         "allow_credentials": True,
         "max_age": 3660,
     },
-    "jwt": {"algorithm": "HS256"},
+    "jwt": {"algorithm": "HS256", "token_expiration": 60 * 60 * 1},
     "commands": {
         "": "guillotina.commands.server.ServerCommand",
         "serve": "guillotina.commands.server.ServerCommand",