+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ CHANGELOG
 
 - PatchField: added operation "multi"
   [masipcat]
+- @duplicate: added option to reset acl
 - Make pytest.mark.app_settings work in older pytest versions too [lferran]
 - @move: destination id conflict should return 409 error, not 412
   [inaki]